create sections and splitted - models, schemas, handlers and saved in the appropriate sections. There are problems with main.py and handlers in every apps section
