+++ apps/authorization/schemas.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class AvailableRolesOut(BaseModel):
+    role_id: int
+    role_name: str
