+++ apps/educational_instuctions/routers.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+router = APIRouter()
