--- schemas/course.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class CourseListOut(BaseModel):
-    course_id: int
-    value: int
