--- schemas/speciality.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-from typing import Dict, Union
-
-
-class SpecialityListOut(BaseModel):
-    university_id: int
-    speciality_id: int
-    speciality_info: Dict[str, Union[int, str]]
